Added support for pitcher_arms MySQL table
Instead of needing to update pitcher's throwing arm every time "pitcher_total_stats" was updated, which took an extremely long time to do, I changed the process. The pitcher's arm used to be stored in the "pitcher_total_stats" MySQL table, but this was inefficient due to the fact a pitcher's arm does not need to be updated like other stats. I created a permanent table to store pitcher arms that can be added to and updated more effectively.